pyproject.toml
[project]
name = "medimind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
app = "backend.main:app"
dependencies = [
    "authlib>=1.7.0",
    "bcrypt>=5.0.0",
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "itsdangerous>=2.2.0",
    "langchain-core>=1.2.23",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.4",
    "mem0ai>=1.0.9",
    "openai>=2.30.0",
    "passlib>=1.7.4",
    "pinecone",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "starlette>=1.0.0",
    "supabase>=2.28.3",
    "uvicorn>=0.42.0",
    "vecs>=0.4.5",
]
